move expunitmiddleware context to model package
