+++ src/main/java/org/breedinginsight/brapps/importer/services/processors/experiment/appendoverwrite/model/ExpUnitMiddlewareContext.java
@@ -1,10 +1,12 @@
-package org.breedinginsight.brapps.importer.services.processors.experiment.model;
+package org.breedinginsight.brapps.importer.services.processors.experiment.appendoverwrite.model;
 
 import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 import org.breedinginsight.brapps.importer.services.processors.experiment.appendoverwrite.model.ExpUnitContext;
 import org.breedinginsight.brapps.importer.services.processors.experiment.create.model.PendingData;
+import org.breedinginsight.brapps.importer.services.processors.experiment.model.ImportContext;
+
 @Getter
 @Setter
 @Builder
